requirements.txt
# Core dependencies
aixplain>=0.2.38
pandas>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0
numpy>=1.24.0

# Web scraping
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Data processing
openpyxl>=3.1.0

# UI options
streamlit>=1.29.0
click>=8.1.0

# API integration
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0

# External integrations
slack-sdk>=3.23.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Utilities
tqdm>=4.66.0
colorama>=0.4.6
